Create trim_adaptors.py
The issue with sequencing reads that are longer than the target fragment is nicely explained in this post on the Genome Factory:
http://thegenomefactory.blogspot.co.uk/2012/11/tools-to-merge-overlapping-paired-end.html

Our PCR amplification design (two-step tailed PCR, with variable Ns to add complexity) impose an additional complication for the available algorithms to deal with. The script above cuts them off.